requirements.txt
numpy==1.20.2
pywin32==227
psd_tools==1.9.17
opencv_python==4.4.0.44
scikit_image==0.16.2
Pillow==8.3.1
skimage==0.0
